refactor: Adjusting code of file
