+++ package.json
@@ -3,51 +3,51 @@
     "version": "0.1.0",
     "private": true,
     "dependencies": {
-      "@fortawesome/fontawesome-svg-core": "^6.5.1",
-      "@fortawesome/free-brands-svg-icons": "^6.5.1",
-      "@fortawesome/free-regular-svg-icons": "^6.5.1",
-      "@fortawesome/free-solid-svg-icons": "^6.5.1",
-      "@fortawesome/react-fontawesome": "^0.2.0",
-      "@testing-library/jest-dom": "^5.17.0",
-      "@testing-library/react": "^13.4.0",
-      "@testing-library/user-event": "^13.5.0",
-      "react": "^18.2.0",
-      "react-dom": "^18.2.0",
-      "react-scripts": "^5.0.1",
-      "sass": "^1.69.5",
-      "web-vitals": "^2.1.4"
+        "@fortawesome/fontawesome-svg-core": "^6.5.1",
+        "@fortawesome/free-brands-svg-icons": "^6.5.1",
+        "@fortawesome/free-regular-svg-icons": "^6.5.1",
+        "@fortawesome/free-solid-svg-icons": "^6.5.1",
+        "@fortawesome/react-fontawesome": "^0.2.0",
+        "@testing-library/jest-dom": "^5.17.0",
+        "@testing-library/react": "^13.4.0",
+        "@testing-library/user-event": "^13.5.0",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "react-scripts": "^5.0.1",
+        "web-vitals": "^2.1.4"
     },
     "scripts": {
-      "start": "react-scripts start",
-      "build": "react-scripts build",
-      "test": "react-scripts test",
-      "eject": "react-scripts eject"
+        "start": "react-scripts start",
+        "build": "react-scripts build",
+        "test": "react-scripts test",
+        "eject": "react-scripts eject"
     },
     "eslintConfig": {
-      "extends": [
-        "react-app",
-        "react-app/jest"
-      ]
+        "extends": [
+            "react-app",
+            "react-app/jest"
+        ]
     },
     "browserslist": {
-      "production": [
-        ">0.2%",
-        "not dead",
-        "not op_mini all"
-      ],
-      "development": [
-        "last 1 chrome version",
-        "last 1 firefox version",
-        "last 1 safari version"
-      ]
+        "production": [
+            ">0.2%",
+            "not dead",
+            "not op_mini all"
+        ],
+        "development": [
+            "last 1 chrome version",
+            "last 1 firefox version",
+            "last 1 safari version"
+        ]
     },
     "devDependencies": {
-      "@babel/core": "^7.23.6",
-      "@babel/preset-env": "^7.23.6",
-      "@babel/preset-react": "^7.23.3",
-      "babel-loader": "^9.1.3",
-      "webpack": "^5.89.0",
-      "webpack-cli": "^5.1.4"
+        "@babel/core": "^7.23.6",
+        "@babel/preset-env": "^7.23.6",
+        "@babel/preset-react": "^7.23.3",
+        "babel-loader": "^9.1.3",
+        "sass": "^1.69.7",
+        "sass-loader": "^13.3.3",
+        "webpack": "^5.89.0",
+        "webpack-cli": "^5.1.4"
     }
-  }
-  
+}
