+++ public/index.html
@@ -4,10 +4,8 @@
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <title>Airnes</title>
-    <link rel="stylesheet" href="styles.css" />
   </head>
   <body>
     <div id="root"></div>
-    <script src="../src/index.js"></script>
   </body>
 </html>
