--- public/styles.css
@@ -1,2 +1,0 @@
-@import url(../src/styles/header/header.css);
-@import url(../src/styles/footer/footer.css);
