+++ src/index.js
@@ -1,7 +1,7 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from 'react-dom/client';
 import App from "./components/App";
 
-const rootElement = document.getElementById('root');
-const root = rootElement.createRoot ? rootElement.createRoot() : ReactDOM;
-root.render(<App />, rootElement);
+const root = document.getElementById('root');
+const rootInstance = createRoot(root);
+rootInstance.render(<App />);
