+++ src/styles/footer.scss
