--- src/styles/header/header.css
@@ -1,140 +1,0 @@
-.header {
-  padding: 10px 10px;
-}
-.header .burger-menu {
-  display: none;
-}
-.header .top {
-  display: flex;
-}
-.header .top .title {
-  flex: 1;
-  display: flex;
-  align-items: center;
-  text-transform: uppercase;
-  font-size: 33px;
-  font-weight: bold;
-  font-family: "Trebuchet MS", "Lucida Sans Unicode", "Lucida Grande", "Lucida Sans", Arial, sans-serif;
-}
-.header .top .icons {
-  display: flex;
-  list-style: none;
-  font-size: 30px;
-}
-.header .top .icons li {
-  margin: 10px 15px;
-}
-.header .top .icons li a {
-  color: black;
-  text-decoration: none;
-}
-.header .navbar {
-  margin-top: 20px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  list-style: none;
-  font-size: 15px;
-  border-top: solid 2px;
-  border-bottom: solid 2px;
-  font-weight: bold;
-  text-transform: uppercase;
-  font-family: "Trebuchet MS", "Lucida Sans Unicode", "Lucida Grande", "Lucida Sans", Arial, sans-serif;
-}
-.header .navbar li {
-  display: block;
-  margin: 10px 15px;
-  padding: 30px 40px;
-  transition: all 0.1s;
-}
-.header .navbar li a {
-  color: black;
-  text-decoration: none;
-}
-.header .navbar li:hover {
-  font-size: 20px;
-  margin: 0px;
-  transition: all 0.1s;
-}
-@media (max-width: 1024px) {
-  .header .header {
-    padding: 30px 50px;
-  }
-  .header .burger-menu {
-    display: block;
-    cursor: pointer;
-    fill: black;
-    transition: fill 0.3s;
-    width: 70px;
-    height: 30px;
-  }
-  .header .title {
-    display: flex;
-    text-transform: uppercase;
-    font-size: 20px;
-    font-weight: bold;
-    font-family: "Trebuchet MS", "Lucida Sans Unicode", "Lucida Grande", "Lucida Sans", Arial, sans-serif;
-  }
-  .header .icon-button {
-    display: flex;
-    list-style: none;
-    font-size: 30px;
-  }
-  .header .navbar {
-    display: none;
-    flex-direction: column;
-    align-items: center;
-    text-align: center;
-    position: absolute;
-    width: 100%;
-    z-index: 1;
-    right: 0;
-  }
-  .header .navbar.show-menu {
-    display: flex;
-  }
-  .header .navbar li {
-    padding: 15px;
-    width: 70%;
-    box-sizing: border-box;
-    border-bottom: 1px solid #555;
-  }
-  .header .navbar li:last-child {
-    border-bottom: none;
-  }
-}
-.header button {
-  border: none;
-  background: none;
-  cursor: pointer;
-}
-.header button svg {
-  fill: black;
-  transition: fill 0.3s;
-  width: 70px;
-  height: 20px;
-}
-.header button svg:hover {
-  fill: #535bf2;
-}
-
-i {
-  position: relative;
-  display: flex !important;
-}
-i .badge {
-  display: block;
-  position: absolute;
-  background-color: red;
-  width: auto;
-  height: auto;
-  right: -5px;
-  bottom: -5px;
-  padding: 3px 7px;
-  color: white;
-  font-size: 10px;
-  font-weight: bold;
-  border-radius: 50px;
-}
-
-/*# sourceMappingURL=header.css.map */
